INFUND-6403 - Adds additional unit tests for CompetitionInviteServiceImpl#getAvailableAssessors
Former-commit-id: 06586f6bf32ea85159a8f5ebca623486757b7dc3

+++ ifs-data-service/data-service-commons/src/main/java/org/innovateuk/ifs/user/repository/UserRepository.java
@@ -51,5 +51,5 @@ public interface UserRepository extends PagingAndSortingRepository<User, Long> {
                                                                    @Param("direction") String direction);
 
     @Query(value = "SELECT COUNT(*) " + AVAILABLE_ASSESSORS_QUERY, nativeQuery = true)
-    long countAllAvailableAssessorsByCompetition(@Param("competitionId") long competitionId);
+    int countAllAvailableAssessorsByCompetition(@Param("competitionId") long competitionId);
 }
